BDO-782: Changed file size type all over to accomodate higher values.

+++ app/controllers/model/PrepareUploadRequestV1.scala
@@ -23,8 +23,8 @@ import services.model.UploadSettings
 
 case class PrepareUploadRequestV1(
   callbackUrl: String,
-  minimumFileSize: Option[Int],
-  maximumFileSize: Option[Int],
+  minimumFileSize: Option[Long],
+  maximumFileSize: Option[Long],
   expectedContentType: Option[String],
   successRedirect: Option[String])
     extends PrepareUpload {
@@ -43,13 +43,13 @@ case class PrepareUploadRequestV1(
 
 object PrepareUploadRequestV1 {
 
-  def reads(maxFileSize: Int): Reads[PrepareUploadRequestV1] =
+  def reads(maxFileSize: Long): Reads[PrepareUploadRequestV1] =
     ((JsPath \ "callbackUrl").read[String] and
-      (JsPath \ "minimumFileSize").readNullable[Int](min(0)) and
-      (JsPath \ "maximumFileSize").readNullable[Int](min(0) keepAnd max(maxFileSize)) and
+      (JsPath \ "minimumFileSize").readNullable[Long](min(0)) and
+      (JsPath \ "maximumFileSize").readNullable[Long](min(0L) keepAnd max(maxFileSize)) and
       (JsPath \ "expectedContentType").readNullable[String] and
       (JsPath \ "successRedirect").readNullable[String])(PrepareUploadRequestV1.apply _)
       .filter(JsonValidationError("Maximum file size must be equal or greater than minimum file size"))(request =>
-        request.minimumFileSize.getOrElse(0) <= request.maximumFileSize.getOrElse(maxFileSize))
+        request.minimumFileSize.getOrElse(0L) <= request.maximumFileSize.getOrElse(maxFileSize))
 
 }
