+++ app/services/model/UploadSettings.scala
@@ -19,8 +19,8 @@ package services.model
 case class UploadSettings(
   uploadUrl: String,
   callbackUrl: String,
-  minimumFileSize: Option[Int],
-  maximumFileSize: Option[Int],
+  minimumFileSize: Option[Long],
+  maximumFileSize: Option[Long],
   expectedContentType: Option[String],
   successRedirect: Option[String],
   errorRedirect: Option[String])
